nerf extremely high strains, re-add angle buff, nerf short maps
a higher value to sectionLength in StarRating() helps nerf maps like a hope in hell, lechenaultia, and other maps that have spaced bursts, but the consistencyMultiplier had to be lowered. the lower consistencyMultiplier made me readd the angle buff, add a short map nerf and buff anglechangebonus in order to balance the system.